quadifyapp/requirements.txt
asn1crypto==0.24.0
bidict==0.22.1
blinker==1.6.3
cbor2==5.4.6
certifi==2018.8.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.4.0

# UPDATED to support OpenSSL 3 on Bookworm
cryptography==41.0.7

entrypoints==0.3
exceptiongroup==1.2.2
h11==0.14.0
idna==2.6
importlib-metadata==6.7.0
iniconfig==2.0.0
keyring==17.1.1
keyrings.alt==3.1.1
luma.core==2.4.2
luma.oled==3.13.0
packaging==24.0
Pillow==9.5.0
pluggy==1.2.0
psutil==5.9.5
pycparser==2.21

# pycrypto CANNOT install on Python 3.11 → replaced with drop-in replacement
pycryptodome==3.20.0

pyftdi==0.54.0
pyserial==3.5
pytest==7.4.4
python-engineio==3.14.2
python-socketio==4.6.1
pyusb==1.2.1
pyxdg==0.25
requests==2.21.0
SecretStorage==2.3.1
simple-websocket==1.1.0
six==1.15.0
smbus2==0.5.0
spidev==3.6
tomli==2.0.1
transitions==0.9.2
typing-extensions==4.7.1
urllib3==1.24.1
websocket-client==1.6.1
wsproto==1.2.0
zipp==3.15.0

CairoSVG>=2.7
cairocffi>=1.6
tinycss2>=1.2
cssselect2>=0.7
defusedxml>=0.7

# EXACTLY AS YOU WROTE IT (just fixed the broken newline)
PyYAML==3.13; python_version < "3.8"
PyYAML==6.0.1; python_version >= "3.8"
